requirements.txt
# Librerías necesarias para el proyecto
discord.py
python-dotenv
pandas